requirements.txt
Django>=5.0
psycopg2-binary>=2.9
djangorestframework>=3.15
djangorestframework-gis>=1.0
django-filter>=24.3
python-dotenv>=1.0
gunicorn>=21.0
requests>=2.31.0
redis>=5.0
